Moved getData function to getData.js, imported getData() to index.js
