+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-    "compile-hero.disable-compile-files-on-did-save-code": true
+    "compile-hero.disable-compile-files-on-did-save-code": false
 }
